fix: moved file to component folder and applied PIGLASAPAT redirect. Also applied changes for readability and redundancy
